requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
drf-spectacular==0.28.0
redis==5.0.1
django-redis==5.4.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
